Sprint-1 POS application completed.

+++ src/main/java/com/ust/pos/user/UserController.java
@@ -1,8 +1,11 @@
 package com.ust.pos.user;
 
 import com.ust.pos.dto.UserDto;
+import com.ust.pos.role.service.RoleService;
 import com.ust.pos.user.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -14,6 +17,9 @@ public class UserController {
     @Autowired
     private UserService userService;
 
+    @Autowired
+    private RoleService roleService;
+
     @GetMapping("/list")
     public String home(Model model) {
         model.addAttribute("users", userService.findAll());
@@ -23,7 +29,10 @@ public class UserController {
     @GetMapping("/get")
     public String update(Model model, @RequestParam String username) {
         UserDto response = userService.findByUserName(username);
-        model.addAttribute("user", response);
+
+        model.addAttribute("userDto", response);
+        model.addAttribute("roles", roleService.findAll());
+
         return "user/user";
     }
 
@@ -31,6 +40,9 @@ public class UserController {
     public String updatePost(Model model, @ModelAttribute UserDto userDto) {
         UserDto response = userService.update(userDto);
         if (!response.isSuccess()) {
+
+            model.addAttribute("userDto", userDto);
+            model.addAttribute("roles", roleService.findAll());
             model.addAttribute("message", response.getMessage());
             return "user/user";
         }
@@ -39,7 +51,20 @@ public class UserController {
 
     @GetMapping("/delete")
     public String delete(Model model, @RequestParam String username) {
-        userService.delete(username);
-        return "user/user";
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (authentication != null) {
+            String loggedInUser = authentication.getName();
+            if (loggedInUser != null) {
+
+                userService.delete(username);
+
+                if (loggedInUser.equals(username)) {
+                    SecurityContextHolder.clearContext();
+                    return "redirect:/login";
+                }
+            }
+        }
+        return "redirect:/user/list";
     }
 }
